refactor: User, Dealer, Player 를 domain 패키지 아래로 이동
