+++ src/main/java/domain/Player.java
@@ -1,4 +1,4 @@
-package domain.user;
+package domain;
 
 public class Player extends User {
