+++ src/main/java/domain/Users.java
@@ -1,8 +1,6 @@
 package domain;
 
-import domain.user.Dealer;
-import domain.user.Player;
-import domain.user.User;
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
