+++ src/test/java/domain/BlackJackTest.java
@@ -2,8 +2,6 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.user.Dealer;
-import domain.user.Player;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
