+++ src/test/java/domain/DealerTest.java
@@ -9,8 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import domain.user.Dealer;
-
 public class DealerTest {
 
 	private Dealer dealer;
