+++ src/test/java/domain/UsersTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.user.Player;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
